fitViewport extends maptiler-sdk
